cli/cmake: Use proper way to find link OpenEXR and Thread libraries.
This fixes the build on Linux distributions that don't put these
libraries in the same path as Ubuntu. Only removing the absolute
paths would be sufficient to fix the build, since CMake is looking
in the system library search paths for the filenames only correctly.
The correct way is to use the CMake function find_package, as done
in this patch.

Since OpenEXR does not ship a working cmake script for find_package,
I am using PkgConfig for the same result, which is available on every
Linux distribution.

This should also be done for the APPLE case, but I didn't have a setup
to test it.